move inv_module_name to ty_inv.rs
